+++ src/components/molecules/Chatbox/Chatbox.scss
@@ -18,7 +18,7 @@
     display: flex;
     flex-flow: column-reverse;
   }
-  
+
   &__messages-infinite-scroll-loading {
     overflow: hidden;
   }
